Overhaul 2 from /home/Code local rewrite of alpha-processing
After much consideration and investigation, I've decided that
the C and C++ build systems will need to be separated.

C++ is just too convoluted a language to use the simpler C
based test binding methods, and since a large part of hanamake's
reason for development was to add C++ support to the build system,
I don't want to half-ass it.

Unfortunately, this means delivery will need to be delayed again.
This pains me when a C-only version could have been done today.

Instead, it may take a month (time permitting).