Fixes mysterious segfault
Essentially, what was happening, was that after
passing all tests on a run, all the log directories
were deleted

Then, changing a only a function in source file A
--
thus only recreating the log directories for
functions in source file A
--
which is depended upon by a function in source file B
--
whose log directory was deleted at the end of the prior
run since it passed all tests
--
in a way that causes both functions in A and B to fail
causes function A to write to its log,
and B to try and write to its log, but the directory
of its log file doesn't exist anymore,
and wasn't recreated
since source file B wasn't updated,
causing fopen("path/to/source file B") to fail,
and a NULL pointer is written to, causing a segfault.

A situation where fopen() failed due to external circumstances
was considered in the past, but I decided then that this was
inherently not the responsibility of hanamake.

In this case, hanamake caused its own failure, and may
have made this debug job trickier than necessary.

POSIX shell script is a powerful and useful tool,
but the complexity of hanamake is now such that a different
tool set may be needed--something more precise than
raw text processing...

Of course, a 3rd reintegration would go a long way toward
alleviating its many gotchas, and I've even considered
renaming the project to 'lit' for LITmus test, since
that's fundamentally what unit and regression tests are anyway,
and I think it's a better name, shorter to type and conceptually
clearer.

I've even got an slight reworking of the command validation script
done, and its way easier to read, benefitting from everything I've
learned while doing this, and some of the more recent cool ideas
for how to really tie everything together, such as the separation
of logs into machine and human versions, and doing basic graphics
rasterization in the style of ncurses using only shell primatives,
taking a page from my time in the unity engine and separating
world state from presentation--presentation is just a function
that reads world state as input, which can be machine log files...

I digress.

Anyway, what's left to do is translating user main functions
to __usermade_main__ throughout the assembly files, adjusting
the C++ parentheses check in .ypp generation to allow a simple
'main' input, and doing a visual touch up translation in
formulating.sh

That would make for a relatively complete beta package and seemless
unit+regression testing suite for C and C++, even being able to test
your main function without missing a beat.

While having full debug mode would be nice, there isn't time for
it right now, which is a little frustrating considering how easy
it would have been to do were this project to only support C code.

But on the bright side, a much wider group of people will be able
to benefit from it if C++ natively supported, especially the students
since C++ is still fairly popular in colleges